scsi: ufs: fix compilation warning on debugfs
msm8994 compilation treats warnings as errors.
function signature should return int instead of ssize_t.

+++ drivers/scsi/ufs/debugfs.c
@@ -600,7 +600,7 @@ static const struct file_operations ufsdbg_dump_device_desc = {
 	.read		= seq_read,
 };
 
-static ssize_t ufsdbg_power_mode_show(struct seq_file *file, void *data)
+static int ufsdbg_power_mode_show(struct seq_file *file, void *data)
 {
 	struct ufs_hba *hba = (struct ufs_hba *)file->private;
 	char *names[] = {
